manifest file exceptionfor setupAirField method line 103

+++ src/com/skilldistillery/jets/JetFighter.java
@@ -6,7 +6,7 @@ public class JetFighter extends Jets implements FighterJet {
 	public JetFighter() {
 	}
 	
-	public JetFighter(String _model, Integer _speed, long _price, int _range, boolean _hasMissiles) {
+	public JetFighter(String _model, double _speed, long _price, int _range, boolean _hasMissiles) {
 		super(_model, _speed, _price, _range);
 		this.hasMissiles = _hasMissiles;
 	}
